cron/news_scrapper_trigger/requirements.txt
beautifulsoup4==4.8.1
Pillow==6.2.1
lxml==4.4.1
requests==2.22.0
redis==3.3.11
selenium==3.141.0
google-cloud-firestore==1.5.0
google-cloud-pubsub==1.0.2
tldextract==2.2.2
